Worked on logic
