+++ app/src/main/java/ru/geekbrains/weatherapplication/fragment/WeatherInfoFragment.java
@@ -15,24 +15,34 @@ import android.widget.Toolbar;
 
 import androidx.annotation.NonNull;
 import androidx.annotation.Nullable;
+import androidx.appcompat.app.AlertDialog;
 import androidx.fragment.app.Fragment;
 import androidx.recyclerview.widget.LinearLayoutManager;
 import androidx.recyclerview.widget.RecyclerView;
 
+import com.google.gson.Gson;
+import com.google.gson.reflect.TypeToken;
+
+import java.io.BufferedReader;
+import java.io.InputStreamReader;
 import java.util.ArrayList;
 import java.util.List;
+import java.util.Observable;
 
 import ru.geekbrains.weatherapplication.R;
 import ru.geekbrains.weatherapplication.adapter.CurrentWeatherExtraAdapter;
 import ru.geekbrains.weatherapplication.adapter.WeatherWeekAdapter;
-import ru.geekbrains.weatherapplication.data.Constants;
 import ru.geekbrains.weatherapplication.data.Parcel;
+import ru.geekbrains.weatherapplication.data.State;
+import ru.geekbrains.weatherapplication.data.dto.CityListItem;
 import ru.geekbrains.weatherapplication.data.request.CurrentWeatherRequest;
 import ru.geekbrains.weatherapplication.data.request.MainRequest;
 import ru.geekbrains.weatherapplication.item.CurrentWeatherExtraItem;
 import ru.geekbrains.weatherapplication.item.OptionItem;
 import ru.geekbrains.weatherapplication.item.WeatherItem;
+import ru.geekbrains.weatherapplication.service.ApiDataReceiver;
 
+import static ru.geekbrains.weatherapplication.data.Constants.CITY_LIST_FILE_PATH;
 import static ru.geekbrains.weatherapplication.data.Constants.LoggerMode.DEBUG;
 import static ru.geekbrains.weatherapplication.data.Constants.WEATHER_OPTIONS;
 
@@ -42,10 +52,13 @@ public class WeatherInfoFragment extends BaseFragment {
 
     private static final String ADDRESS_WEATHER = "https://www.gismeteo.ru/";
 
+    private View mainFragment;
+    private View viewLoading;
+
     private Toolbar toolbar;
     private TextView tempTextView;
 
-    private ImageView image_weather;
+    private ImageView imageWeather;
 
     private CurrentWeatherExtraAdapter extraInfoAdapter;
     private RecyclerView extraInfoRecycler;
@@ -55,25 +68,12 @@ public class WeatherInfoFragment extends BaseFragment {
 
     private Button btnMoreInfo;
 
+    private String cityName;
 
-    public static WeatherInfoFragment newInstance(String cityName, List<OptionItem> options) {
+    public static WeatherInfoFragment newInstance(String cityName) {
         WeatherInfoFragment fragment = new WeatherInfoFragment();
         Bundle args = new Bundle();
-        args.putSerializable(WEATHER_OPTIONS, new Parcel(cityName, options));
-        fragment.setArguments(args);
-        return fragment;
-    }
-
-    public static WeatherInfoFragment newInstance(CurrentWeatherRequest weatherRequest) {
-        WeatherInfoFragment fragment = new WeatherInfoFragment();
-        Bundle args = new Bundle();
-
-        String icon = "00d";
-        if (weatherRequest.getWeather() != null) {
-            icon = weatherRequest.getWeather()[0].getIcon();
-        }
-
-        args.putSerializable(WEATHER_OPTIONS, new Parcel(weatherRequest.getName(), new ArrayList<>(), icon));
+        args.putSerializable(WEATHER_OPTIONS, new Parcel(cityName, new ArrayList<>()));
         fragment.setArguments(args);
         return fragment;
     }
@@ -95,21 +95,24 @@ public class WeatherInfoFragment extends BaseFragment {
             parcel = (Parcel) savedInstanceState.getSerializable(WEATHER_OPTIONS);
         }
 
-        if (parcel == null) {
-            Toast.makeText(getContext(), getString(R.string.cannot_load_info_error), Toast.LENGTH_LONG);
-        }
-        else {
-            String title = getString(R.string.weather_info_title, parcel.cityName);
-            if (DEBUG) {
-                Log.d("WeatherInfoFragment", "title = " + title);
-            }
-            toolbar.setTitle(title);
+        cityName = parcel.cityName;
+        loadWeatherData(findCityByName(parcel.cityName));
 
-            image_weather.setImageResource(Constants.getWeatherImage(parcel.icon));
-
-            setupRecycler(view, parcel.options);
-            btnMoreInfo.setEnabled(true);
-        }
+//        if (parcel == null) {
+//            Toast.makeText(getContext(), getString(R.string.cannot_load_info_error), Toast.LENGTH_LONG);
+//        }
+//        else {
+//            String title = getString(R.string.weather_info_title, parcel.cityName);
+//            if (DEBUG) {
+//                Log.d("WeatherInfoFragment", "title = " + title);
+//            }
+//            toolbar.setTitle(title);
+//
+//            imageWeather.setImageResource(Constants.getWeatherImage(parcel.icon));
+//
+//            setupRecycler(view, parcel.options);
+//            btnMoreInfo.setEnabled(true);
+//        }
 
     }
 
@@ -119,11 +122,14 @@ public class WeatherInfoFragment extends BaseFragment {
     }
 
     private void bindView(View view) {
+        mainFragment = view.findViewById(R.id.weather_day_info);
+        viewLoading = view.findViewById(R.id.frame_loading);
+
         toolbar = getActivity().findViewById(R.id.toolbar);
 
         tempTextView = getActivity().findViewById(R.id.current_temp);
 
-        image_weather = getActivity().findViewById(R.id.image_current_weather);
+        imageWeather = getActivity().findViewById(R.id.image_current_weather);
 
         weatherWeekRecycler = view.findViewById(R.id.weather_week_recycler);
         weatherWeekRecycler.setVisibility(View.GONE);
@@ -190,6 +196,97 @@ public class WeatherInfoFragment extends BaseFragment {
 
     @Override
     void updateView(MainRequest data) {
+
+    }
+
+    private void loadWeatherData(CityListItem city) {
+        showStateView(State.Loading);
+
+        if (city != null) {
+            if (DEBUG) {
+                Log.d(TAG, "found city:" + city);
+            }
+            getWeather(city, 1);
+        }
+        else {
+            if (DEBUG) {
+                Log.e(TAG, "Not found city");
+            }
+
+            showStateView(State.ErrorData);
+        }
+
+    }
+
+    private CityListItem findCityByName(String cityName) {
+        try {
+            BufferedReader reader = new BufferedReader(new InputStreamReader(getContext().getAssets().open(CITY_LIST_FILE_PATH)));
+
+            List<CityListItem> cities = new Gson().fromJson(reader, new TypeToken<List<CityListItem>>() {}.getType());
+
+            CityListItem res = cities.stream()
+                    .filter(city -> cityName.equals(city.getName()))
+                    .findAny().orElse(null);
+
+            reader.close();
+            return res;
+        }
+        catch (Exception ex) {
+            ex.printStackTrace();
+            return null;
+        }
+    }
+
+    public void getWeather(CityListItem city, int requestMode) {
+        new Thread(new ApiDataReceiver(this, this, city, requestMode)).start();
+    }
+
+    private void showStateView(@NonNull State state){
+        switch (state){
+            case HasData:
+                viewLoading.setVisibility(View.GONE);
+
+                mainFragment.setVisibility(View.VISIBLE);
+                break;
+
+            case Loading:
+                mainFragment.setVisibility(View.GONE);
+
+                viewLoading.setVisibility(View.VISIBLE);
+                break;
+
+            case ErrorData:
+                viewLoading.setVisibility(View.GONE);
+                mainFragment.setVisibility(View.GONE);
+
+                createAlertDialog(getText(R.string.cannot_load_info_error));
+                break;
+
+            default:
+                throw new IllegalArgumentException("Wrong state: " + state);
+        }
+
+    }
+
+    private void createAlertDialog(CharSequence errorMsg) {
+        AlertDialog.Builder builder = new AlertDialog.Builder(getContext());
+        builder.setTitle(R.string.error_title)
+                .setMessage(errorMsg)
+                .setIcon(R.drawable.ic_zip)
+                .setCancelable(true)
+                .setNegativeButton(R.string.button_retry,
+                        (dialog, id) -> loadWeatherData(findCityByName(cityName)))
+                .setPositiveButton(R.string.button_back, (dialog, id) -> getActivity().getSupportFragmentManager().popBackStack());
+    }
+
+    @Override
+    public void update(Observable observable, Object o) {
+        if (o instanceof ApiDataReceiver) {
+            ApiDataReceiver data = (ApiDataReceiver) o;
+
+
+
+        }
 
     }
 }
