+++ app/src/main/res/values-ru-rRU/strings.xml
@@ -11,6 +11,7 @@
     <string name="show_forecast_yes">Вперед</string>
     <string name="button_ok">ОК</string>
     <string name="button_retry">Попробовать снова</string>
+    <string name="button_back">Назад</string>
 
     <string name="weather_info_title">Погода в г.%1$s</string>
